Implements generic style for all filterbars
