+++ layouts/partials/card/container.html
@@ -1,13 +1,18 @@
 {{ $data := (index $.Site.Data .Page.Params.dataSrc ) }}
 <div
   id="generic-filter-test"
-  class="filterable"
+  class="card-container"
   x-data="RMST_TS.createFilterBarInstance({{ $data.config | jsonify }})"
 >
-  {{ partial "filterbar/bar" . }}
+  <aside
+    class="card-container-sidebar"
+    x-intersect:enter="$store.mobileNavigationStore.unregisterBackToTop()"
+    x-intersect:leave="$store.mobileNavigationStore.registerBackToTop(RMST_TS.createScrollToElement($el))"
+  >
+    {{ partial "filterbar/bar" . }}
+  </aside>
 
-
-  <div class="item-container">
+  <div class="card-container-main">
     {{ $cardLayout := $data.config.cardLayout }}
     {{ $classes := $data.config.itemClasses }}
     {{ $i18nPrefix := index $data.config "i18nPrefix" }}
